Remove no longer useful update-properties tool (#184)
We are running automatic updates with dependabot. It takes care of all
sub-modules correctly, but skips the parent configuration which is only
used for the 'updates-properties.groovy' script, forcing a manual update
for those.
So we are removing it since now it creates work more than helps.
Manual update is still possible just using any "replace all" tool.